backend/requirements.txt
asyncpg
fastapi
passlib[bcrypt]
psycopg2-binary
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
sqlalchemy
uvicorn[standard]
python-multipart